Share QR login device copy helpers
Move the duplicated signed-in headline and device subline builders into one helper module so the success step and consumed panel keep the same fallback behavior and translation strings.
